Update the webview service to use the new service pattern
