+++ src/renderer/components/basic-list/basic-list.component.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { SavedTabInfo, TabInfo } from '@shared/data/web-view.model';
+import { SavedTabInfo, TabInfo } from '@shared/models/docking-framework.model';
 import {
   ExpandedState,
   useReactTable,
