+++ src/renderer/components/projects/download-update-project-tab.component.tsx
@@ -1,4 +1,4 @@
-import { SavedTabInfo, TabInfo } from '@shared/data/web-view.model';
+import { SavedTabInfo, TabInfo } from '@shared/models/docking-framework.model';
 import {
   List,
   ListItem,
