+++ src/renderer/components/settings-dialog/settings-tab.component.tsx
@@ -9,7 +9,7 @@ import {
   Tabs,
   Typography,
 } from '@mui/material';
-import { SavedTabInfo, TabInfo } from '@shared/data/web-view.model';
+import { SavedTabInfo, TabInfo } from '@shared/models/docking-framework.model';
 import {
   SyntheticEvent,
   useState,
