+++ src/shared/models/web-view.model.ts
@@ -1,63 +1,4 @@
-import { Dispatch, ReactNode, SetStateAction } from 'react';
-
-/**
- * Saved information used to recreate a tab.
- *
- * - {@link TabLoader} loads this into {@link TabInfo}
- * - {@link TabSaver} saves {@link TabInfo} into this
- */
-export type SavedTabInfo = {
-  /**
-   * Tab ID - a unique identifier that identifies this tab. If this tab is a WebView, this ID will
-   * match the `WebViewDefinition.id`
-   */
-  id: string;
-  /** Type of tab - indicates what kind of built-in tab this info represents */
-  tabType: string;
-  /** Data needed to load the tab */
-  data?: unknown;
-};
-
-/**
- * Information that Paranext uses to create a tab in the dock layout.
- *
- * - {@link TabLoader} loads {@link SavedTabInfo} into this
- * - {@link TabSaver} saves this into {@link SavedTabInfo}
- */
-export type TabInfo = SavedTabInfo & {
-  /**
-   * Url of image to show on the title bar of the tab
-   *
-   * Defaults to Platform.Bible logo
-   */
-  tabIconUrl?: string;
-  /** Text to show on the title bar of the tab */
-  tabTitle: string;
-  /** Content to show inside the tab. */
-  content: ReactNode;
-  /** (optional) Minimum width that the tab can become in CSS `px` units */
-  minWidth?: number;
-  /** (optional) Minimum height that the tab can become in CSS `px` units */
-  minHeight?: number;
-};
-
-/**
- * Function that takes a {@link SavedTabInfo} and creates a Paranext tab out of it. Each type of tab
- * must provide a {@link TabLoader}.
- *
- * For now all tab creators must do their own data type verification
- */
-export type TabLoader = (savedTabInfo: SavedTabInfo) => TabInfo;
-
-/**
- * Function that takes a Paranext tab and creates a saved tab out of it. Each type of tab can
- * provide a {@link TabSaver}. If they do not provide one, the properties added by `TabInfo` are
- * stripped from TabInfo by `saveTabInfoBase` before saving (so it is just a {@link SavedTabInfo}).
- *
- * @param tabInfo The Paranext tab to save
- * @returns The saved tab info for Paranext to persist. If `undefined`, does not save the tab
- */
-export type TabSaver = (tabInfo: TabInfo) => SavedTabInfo | undefined;
+import { Dispatch, SetStateAction } from 'react';
 
 /** The type of code that defines a webview's content */
 export enum WebViewContentType {
@@ -225,9 +166,6 @@ export type SavedWebViewDefinition =
     ) &
       Pick<WebViewDefinitionBase, 'id' | 'webViewType'>;
 
-/** Props that are passed to the web view tab component */
-export type WebViewTabProps = WebViewDefinition;
-
 /** The properties on a WebViewDefinition that may be updated when that webview is already displayed */
 // To allow more properties to be updated, add them in
 // `web-view.service.ts` -> `getUpdatablePropertiesFromWebViewDefinition`
@@ -326,62 +264,8 @@ export type WebViewProps = {
   updateWebViewDefinition: UpdateWebViewDefinition;
 };
 
-/** Information about a tab in a panel */
-interface TabLayout {
-  type: 'tab';
-}
-
-/**
- * Indicates where to display a floating window
- *
- * - `cascade` - place the window a bit below and to the right of the previously created floating
- *   window
- * - `center` - center the window in the dock layout
- */
-type FloatPosition = 'cascade' | 'center';
-
-/** The dimensions for a floating tab in CSS `px` units */
-export type FloatSize = { width: number; height: number };
-
-/** Information about a floating window */
-export interface FloatLayout {
-  type: 'float';
-  floatSize?: FloatSize;
-  /** Where to display the floating window. Defaults to `cascade` */
-  position?: FloatPosition;
-}
-
-export type PanelDirection =
-  | 'left'
-  | 'right'
-  | 'bottom'
-  | 'top'
-  // TODO: The following produce a panel but probably aren't useful for panels - IJH 2023-05-5
-  | 'before-tab'
-  | 'after-tab'
-  | 'maximize'
-  | 'move'
-  | 'active'
-  | 'update';
-
-/** Information about a panel */
-interface PanelLayout {
-  type: 'panel';
-  direction?: PanelDirection;
-  /** If undefined, it will add in the `direction` relative to the previously added tab. */
-  targetTabId?: string;
-}
-
-/** Information about how a Paranext tab fits into the dock layout */
-export type Layout = TabLayout | FloatLayout | PanelLayout;
-
-/** Event emitted when webViews are created */
-export type AddWebViewEvent = {
-  webView: SavedWebViewDefinition;
-  layout: Layout;
-};
-
 /** Options that affect what `webViews.getWebView` does */
+// This can't live in web-view.service-model.ts because extensions can see it
 export type GetWebViewOptions = {
   /**
    * If provided and if a web view with this ID exists, requests from the web view provider an
